fix : implemented protected routes auth
