+++ src/app/homepage/links/layout.tsx
@@ -1,8 +1,9 @@
-import Header from "../../components/UI/Header";
+import Header from "@/components/UI/Header";
+
 const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   return (
     <div>
-      <Header />
+      {/* <Header /> */}
       <div>{children}</div>
     </div>
   );
