+++ src/app/homepage/profile/layout.tsx
@@ -1,0 +1,5 @@
+const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+  return <div>{children}</div>;
+};
+
+export default Layout;
