+++ src/app/preview/page.tsx
@@ -1,61 +1,12 @@
-import { Button } from "@nextui-org/react";
-import { Avatar } from "antd";
-import React from "react";
-import {
-  FaFreeCodeCamp,
-  FaGithub,
-  FaLinkedin,
-  FaYoutube,
-} from "react-icons/fa";
-import ButtonLink from "./ButtonLink";
-import { SiCodewars } from "react-icons/si";
-import Header from "./Header";
+"use client";
+import Preview from "./Preview";
 
 type Props = {};
 
 const page = (props: Props) => {
   return (
-    <div className="flex flex-col items-center justify-center p-4">
-      <Header />
-
-      <main className="z-10 mx-auto mt-36 flex flex-col items-center overflow-y-auto py-3 md:h-[560px] md:w-[350px] md:rounded-xl md:bg-white md:shadow-md">
-        <div className="mb-14 flex flex-col gap-y-6">
-          <Avatar
-            shape="circle"
-            size={150}
-            className="border-4 border-primary"
-            alt="Profile Picture"
-            src="/eminem.jpeg"
-          />
-          <p className="text-3xl font-bold">John Doe</p>
-          <span className="text-grey">example@gmail.com</span>
-        </div>
-
-        <div className="flex flex-col gap-y-5">
-          <ButtonLink bgColor="black" text="Github" leftIcon={<FaGithub />} />
-          <ButtonLink
-            bgColor="#EE3939"
-            text="Youtube"
-            leftIcon={<FaYoutube />}
-          />
-          <ButtonLink
-            bgColor="#2D68FF"
-            text="LinkedIn"
-            leftIcon={<FaLinkedin />}
-          />
-          <ButtonLink
-            bgColor="#8A1A50"
-            text="Codewars"
-            leftIcon={<SiCodewars />}
-          />
-          <ButtonLink
-            bgColor="#302267"
-            text="freeCodeCamp"
-            leftIcon={<FaFreeCodeCamp />}
-          />
-        </div>
-      </main>
-      <div className="file absolute left-0 right-0 top-0 hidden h-[357px] rounded-b-[32px] bg-primary md:block"></div>
+    <div>
+      <Preview />
     </div>
   );
 };
