--- src/app/profile/layout.tsx
@@ -1,11 +1,0 @@
-import Header from "../../components/UI/Header"
-const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-    return (
-        <div>
-            <Header />
-            <div>{children}</div>
-        </div>
-    );
-};
-
-export default Layout;
