+++ src/components/header/HeaderLinks.tsx
@@ -1,47 +1,47 @@
-"use client"
+"use client";
 import Link from "next/link";
-import { CgProfile } from "react-icons/cg";
 import { FiLink } from "react-icons/fi";
 import clsx from "clsx";
 import { usePathname } from "next/navigation";
 import { FaUser } from "react-icons/fa6";
 
 const HeaderLinks: React.FC = () => {
-    const currentPath = usePathname()
-    //   const currentRoute = router.pathname;
+  const currentPath = usePathname();
 
-    const isActiveRoute = (baseRoute: string) =>
-        currentPath.startsWith(baseRoute);
+  const isActiveRoute = (baseRoute: string) =>
+    currentPath.startsWith(baseRoute);
 
-    return (
-        <nav className="flex flex-row gap-2">
-            <Link
-                href="/links"
-                className={clsx(
-                    "rounded-xl px-5 py-[11px] hover:text-primary flex items-center gap-x-1",
-                    isActiveRoute("/links")
-                        ? "bg-light-purple text-primary"
-                        : "bg-white text-grey"
-                )}
-            >
-                <FiLink height={15.63} width={15.63} />
-                <span className="font-semibold hidden sm:inline-block">Links</span>
-            </Link>
+  return (
+    <nav className="flex flex-row gap-2">
+      <Link
+        href="/homepage/links"
+        className={clsx(
+          "flex items-center gap-x-1 rounded-xl px-5 py-[11px] hover:text-primary",
+          isActiveRoute("/homepage/links")
+            ? "bg-light-purple text-primary"
+            : "bg-white text-grey",
+        )}
+      >
+        <FiLink height={15.63} width={15.63} />
+        <span className="hidden font-semibold sm:inline-block">Links</span>
+      </Link>
 
-            <Link
-                href="/profile"
-                className={clsx(
-                    "rounded-xl px-5 py-[11px] hover:text-primary flex items-center gap-x-1",
-                    isActiveRoute("/profile")
-                        ? "bg-light-purple text-primary"
-                        : "bg-white text-grey"
-                )}
-            >
-                <FaUser height={15.63} width={15.63} />
-                <span className="font-semibold capitalize hidden sm:inline-block">Profile Details</span>
-            </Link>
-        </nav>
-    );
+      <Link
+        href="/homepage/profile"
+        className={clsx(
+          "flex items-center gap-x-1 rounded-xl px-5 py-[11px] hover:text-primary",
+          isActiveRoute("/homepage/profile")
+            ? "bg-light-purple text-primary"
+            : "bg-white text-grey",
+        )}
+      >
+        <FaUser height={15.63} width={15.63} />
+        <span className="hidden font-semibold capitalize sm:inline-block">
+          Profile Details
+        </span>
+      </Link>
+    </nav>
+  );
 };
 
 export default HeaderLinks;
